always delete saves from fs + split delete dialogs
