+++ frontend/src/types/emitter.d.ts
@@ -75,7 +75,6 @@ export type Events = {
   firmwareDrawerShow: null;
   updateDataTablePages: null;
   sortBarShow: null;
-  romUpdated: DetailedRom;
   showQRCodeDialog: SimpleRom;
   selectSaveDialog: DetailedRom;
   selectStateDialog: DetailedRom;
